+++ test/mri/excludes/TestKeywordArgumentsSymProcRefinements.rb
@@ -1,4 +1,3 @@
-exclude :test_splat_fixnum, "needs investigation"
 exclude :test_sym_proc_refine_kwsplat, "needs investigation"
 exclude :test_sym_proc_refine_method_missing_kwsplat, "needs investigation"
 exclude :test_sym_proc_refine_super_method_missing_kwsplat, "needs investigation"